Refactor variable names and move Point
