+++ src/main/java/org/team114/lib/pathgenerator/Path.java
@@ -3,7 +3,7 @@ package org.team114.lib.pathgenerator;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.team114.lib.util.Point;
+import org.team114.lib.geometry.Point;
 
 /**
  * An abstract piecewise path section.
@@ -23,9 +23,10 @@ public class Path {
 
     /**
      * Creates a path based on pre-generated lists of of the x and y spline components. This method
-     * is not recommended through and if possible the PathFactory should be used.
+     * is not recommended however and if possible the PathFactory should be used.
      * @param xComponents to use
      * @param yComponents to use
+     * @see PathFactory
      */
     public Path(List<PolynomialSpline> xComponents, List<PolynomialSpline> yComponents) {
         assert(xComponents.size() == yComponents.size());
@@ -38,7 +39,7 @@ public class Path {
      * Creates a path based on a the long list of coefficients given by the spline viewer application.
      * Array is ordered as [piecewise section][x component or y component][coefficient]. It is not
      * recommended to use this method if not copy pasting from the spline viewer application.
-     * @param coefficients
+     * @param coefficients coefficients from the spline viewer
      */
     public Path(double[][][] coefficients) {
         xComponents = new ArrayList<PolynomialSpline>();
@@ -52,7 +53,8 @@ public class Path {
 
     /**
      * Gets the point on path at t.
-     * @param t is how far along the path to get the point.
+     * @param t is how far along the path to get the point, measured from n through n+1 for each spline
+     *          segment, beginning with the 0th.
      * @return The Point at t.
      */
     public Point getPointAtT(double t) {
@@ -63,10 +65,10 @@ public class Path {
         if(t == length()) { //Prevent out of bounds
             t = 1;
             component = (int) (length() - 1);
-        }else {
-            t = t - ((int) t);
+        } else {
+            t = t - component;
         }
-        return new Point(xComponents.get(component).eval(t), yComponents.get(component).eval(t));
+        return new Point(xComponents.get(component).getValueAt(t), yComponents.get(component).getValueAt(t));
     }
 
     /**
@@ -74,7 +76,7 @@ public class Path {
      * @param t is how far along the path to get the point.
      * @return The Point containing the x and y derivatives.
      */
-    public Point dfdt(double t) {
+    public Point dydx(double t) {
         if (t < 0 || t > length()) {
             throw new IndexOutOfBoundsException("The parameter must be between 0 and " + length() + ": " + t);
         }
@@ -82,7 +84,7 @@ public class Path {
         if(t == length()) { //Prevent out of bounds
             t = 1;
             component = (int) (length() - 1);
-        }else {
+        } else {
             t = t - ((int) t);
         }
         return new Point(xComponents.get(component).dfdt(t), yComponents.get(component).dfdt(t));
@@ -93,31 +95,24 @@ public class Path {
      * @return The spline length.
      */
     public double length() {
+        assert(xComponents.size() == yComponents.size());
         return xComponents.size();
     }
     
     /**
-     * Gets a point along a normal line based at t down the spline. If n is positive then the
-     * normal line is used which is on the right side of the direction of the derivative. If
-     * n is negative, the other side is used.
-     * @param t is the base of the normal line.
-     * @param n is the length of the normal line.
+     * Gets a point along a normal line based at t down the spline. If n is negative the point will be
+     * 90 degrees clockwise of the direction of the derivative line. If positive, it will be 90 degrees
+     * counterclockwise. This can also be interpreted as right (when facing towards increasing t) if positive
+     * and left when negative.
+     * @param t is the location on the Path to extrapolate from.
+     * @param n is the distance from the point to the path along the normal line.
      * @return The point along the normal line
      */
     public Point getPointAlongNormal(double t, double n) {
         Point base = getPointAtT(t);
-        Point dfdt = dfdt(t);
-        
-        if(dfdt.x() == 0 && dfdt.y() == 0)
-            throw new RuntimeException("Can not find a normal line from a derivative of 0.");
-        
-        double h = Math.sqrt(dfdt.x() * dfdt.x() + dfdt.y() * dfdt.y());
-        double angle = Math.acos(dfdt.x() / h);
-        
-        if(Math.asin(dfdt.y() / h) < 0) //correct angle
-            angle = 2 * Math.PI - angle;
-        angle -= Math.PI / 2; // add 90 degrees
-        
+        Point dydx = dydx(t);
+        // atan2 handles infinite or 0 derivatives
+        double angle = Math.atan2(dydx.y(), dydx.x()) - Math.PI/2;
         return new Point(base.x() + n * Math.cos(angle), base.y() + n * Math.sin(angle));
     }
 
@@ -127,18 +122,18 @@ public class Path {
      * segments. If in the event points are essential, a wrapper may be useful.
      * @return A list of all of the end points of the parametric components.
      */
-    public List<Point> generatePointList(){
+    public List<Point> generatePointList() {
         ArrayList<Point> points = new ArrayList<Point>();
-        for(int i = 0; i <= xComponents.size(); i++)
+        for(int i = 0; i <= length(); i++)
             points.add(getPointAtT(i));
         return points;
     }
 
-    public List<PolynomialSpline> getX(){
+    public List<PolynomialSpline> getXList(){
         return xComponents;
     }
 
-    public List<PolynomialSpline> getY(){
+    public List<PolynomialSpline> getYList(){
         return yComponents;
     }
 }
